+++ docs/versionhistory.rst
@@ -1,13 +1,14 @@
 Version history
 ===============
 
-This library adheres to `Semantic Versioning <http://semver.org/>`_.
+This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
 **UNRELEASED**
 
 - **BACKWARDS INCOMPATIBLE** Switched Sentry reporter to use sentry-sdk instead of raven
 - Added support for Python 3.10
 - Dropped support for Python 3.5 and 3.6
+- Fixed error in default exception handler when the context lacks the ``exception`` key
 
 **1.0.0** (2017-11-26)