+++ docs/versionhistory.rst
@@ -3,7 +3,7 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 
-**UNRELEASED**
+**2.0.0** (2022-03-29)
 
 - **BACKWARDS INCOMPATIBLE** Switched Sentry reporter to use sentry-sdk instead of raven
 - Added support for Python 3.10